Drop redundant fp32 logits cast; restore num_blocks=9 for SDPA-fp32 test
The .float() cast on lm_head output was redundant — F.cross_entropy
upcasts softmax to fp32 internally even with bf16 logits. The 2 GB
saved-for-backward fp32 logits tensor is what was OOM'ing at 14.96 GiB
allocation (saved tensor for the 262K-vocab × 2048-seq × fp32 softmax
backward).

With the perceiver SDPA-fp32 fix landed, this redundant cast can go.
Numerical stability for the lm_head softmax is owned by torch's CE.

Restore num_blocks=9 in the YAML — the actual test of whether the
perceiver SDPA fp32 fix lets us run at depth.

+++ src/ctx_to_lora/modeling/hypernet.py
@@ -826,12 +826,13 @@ class ModulatedPretrainedModel(nn.Module):
         ):
             from transformers.modeling_outputs import CausalLMOutputWithPast
 
-            # Cast logits to fp32: CE over 262K Gemma 4 classes is unstable
-            # in bf16 (softmax overflow → NaN grads). The loss layer takes
-            # fp32 anyway, so this is the standard upcast.
-            logits = self.base_model.lm_head(
-                model_outputs.last_hidden_state
-            ).float()
+            # Keep logits in the input dtype (bf16). F.cross_entropy upcasts
+            # softmax to fp32 internally, so the manual .float() cast we used
+            # to do was redundant — and at 262K vocab × 2048 packed seq, the
+            # fp32 logits tensor is ~2 GB; the saved tensor for backward
+            # doubled that. With the perceiver SDPA fp32 fix in place, the
+            # numerical concern is handled there.
+            logits = self.base_model.lm_head(model_outputs.last_hidden_state)
             model_outputs = CausalLMOutputWithPast(
                 loss=None,
                 logits=logits,
